Implementação do CRUD

+++ src/conta/model/ContaPoupanca.java
@@ -1,14 +1,14 @@
 package conta.model;
 
 public class ContaPoupanca extends Conta{
-	private int aniversario;
+	private float aniversario;
 
-	public ContaPoupanca(int numero, int agencia, int tipo, String titular, float saldo, int aniversario) {
+	public ContaPoupanca(int numero, int agencia, int tipo, String titular, float saldo, float aniversario) {
 		super(numero, agencia, tipo, titular, saldo);
 		this.aniversario = aniversario;
 	}
 
-	public int getAniversario() {
+	public float getAniversario() {
 		return aniversario;
 	}
